refactor(internal): split deep-search mixin into scoring/fetch/scene layers
Addresses review feedback that _deep.py (1343 lines) exceeded the ~1000-line
guideline this PR adds. Extract three single-concern mixins into a linear
chain so each module stays well under the threshold:

  ScoringMixin (_scoring) <- ConfigFetchMixin (_fetch)
    <- SceneSearchMixin (_scenes) <- DeepSearchMixin (_deep)

- _scoring: config-dict relevance scoring (BM25 fuzzy + exact substring)
- _fetch: shared 3-tier config fetch + entry scoring (used by deep + scene)
- _scenes: scene registry walk + per-id fetch (scenes lack a list primitive)
- _deep drops 1343 -> 767 lines; largest module now well under ~1000
- Behavior-preserving (verbatim method moves); AUTOMATION_CONFIG_TIME_BUDGET
  patch target stays in _deep; AGENTS.md example de-enumerated
