+++ src/ha_mcp/tools/smart_search/_deep.py
@@ -3,35 +3,29 @@
 import asyncio
 import logging
 import re
-import time
-from collections.abc import Awaitable, Callable
 from typing import Any
 
 from fastmcp import Context
 from fastmcp.exceptions import ToolError
 
-from ...utils.fuzzy_search import BM25Scorer, calculate_ratio, tokenize
 from ..helpers import exception_to_structured_error, safe_info, safe_progress
 from ..tools_config_dashboards import fetch_dashboards_list
 from ..tools_config_entry_flow import FLOW_HELPER_TYPES
 from ..tools_integrations import fetch_entry_options
-from ._base import _SearchBase
 from ._config import (
     AUTOMATION_CONFIG_TIME_BUDGET,
     BULK_REST_TIMEOUT,
-    BULK_WEBSOCKET_TIMEOUT,
     DEFAULT_CONCURRENCY_LIMIT,
     INDIVIDUAL_CONFIG_TIMEOUT,
-    INDIVIDUAL_FETCH_BATCH_SIZE,
-    SCENE_CONFIG_TIME_BUDGET,
     SCRIPT_CONFIG_TIME_BUDGET,
 )
+from ._scenes import SceneSearchMixin
 
 logger = logging.getLogger(__name__)
 
 
-class DeepSearchMixin(_SearchBase):
-    """``deep_search`` and its config-fetch/scoring helpers."""
+class DeepSearchMixin(SceneSearchMixin):
+    """deep_search orchestration + per-type automation/script/helper/dashboard/flow search."""
 
     async def deep_search(
         self,
@@ -224,140 +218,6 @@ class DeepSearchMixin(_SearchBase):
                     uid_map[eid] = uid
         return uid_map
 
-    @staticmethod
-    def _index_configs(
-        items: list[dict[str, Any]],
-        id_of: Callable[[dict[str, Any]], str | None],
-    ) -> dict[str, dict[str, Any]]:
-        """Build a ``{id: config}`` map, skipping items with no usable id."""
-        configs: dict[str, dict[str, Any]] = {}
-        for item in items:
-            key = id_of(item)
-            if key:
-                configs[key] = item
-        return configs
-
-    async def _bulk_fetch_configs(
-        self,
-        rest_endpoint: str,
-        ws_types: list[str],
-        id_of: Callable[[dict[str, Any]], str | None],
-        rest_timeout: float,
-        label: str,
-    ) -> dict[str, dict[str, Any]] | None:
-        """Bulk-fetch all configs of one domain: REST endpoint, then WS list endpoints.
-
-        Returns ``{id: config}`` (possibly empty) on the first successful
-        attempt, or ``None`` when every attempt failed. An empty-but-successful
-        REST list returns ``{}`` (not ``None``) so the caller skips the
-        individual-fetch fallback exactly as it would for a populated response.
-        """
-        try:
-            resp = await asyncio.wait_for(
-                self.client._request("GET", rest_endpoint),
-                timeout=rest_timeout,
-            )
-            if isinstance(resp, list):
-                return self._index_configs(resp, id_of)
-        except Exception as e:
-            logger.debug(f"{label} REST bulk fetch failed: {e}")
-
-        for ws_type in ws_types:
-            try:
-                ws_resp = await asyncio.wait_for(
-                    self.client.send_websocket_message({"type": ws_type}),
-                    timeout=BULK_WEBSOCKET_TIMEOUT,
-                )
-                if isinstance(ws_resp, dict) and ws_resp.get("success"):
-                    return self._index_configs(ws_resp.get("result", []), id_of)
-            except Exception as e:
-                logger.debug(f"{label} WebSocket bulk fetch ({ws_type}) failed: {e}")
-        return None
-
-    async def _individual_fetch_budgeted(
-        self,
-        ids: list[str],
-        fetch_one: Callable[[str], Awaitable[tuple[str, dict[str, Any] | None]]],
-        budget: float,
-        label: str,
-        plural: str,
-    ) -> tuple[dict[str, dict[str, Any]], int, int]:
-        """Fetch configs individually in parallel batches under a wall-clock budget.
-
-        ``fetch_one(id)`` returns ``(id, config | None)``. New batches stop
-        launching once ``budget`` seconds elapse. Returns
-        ``(configs, failed_count, skipped_count)``.
-
-        Fetch order is NOT prioritized by name score: deep_search's purpose is
-        to find matches INSIDE configs (conditions/actions), not just by name,
-        so name-prioritizing would skip the configs most likely to contain
-        non-obvious matches. See #879.
-        """
-        configs: dict[str, dict[str, Any]] = {}
-        budget_start = time.perf_counter()
-        total_to_fetch = len(ids)
-        fetched_count = 0
-        failed_count = 0
-        skipped_count = 0
-        for i in range(0, len(ids), INDIVIDUAL_FETCH_BATCH_SIZE):
-            if time.perf_counter() - budget_start > budget:
-                skipped_count = total_to_fetch - fetched_count - failed_count
-                logger.warning(
-                    f"{label} config fetch budget exhausted ({budget}s). "
-                    f"Fetched {fetched_count}/{total_to_fetch} "
-                    f"({failed_count} failed), skipped {skipped_count} {plural}."
-                )
-                break
-            batch = ids[i : i + INDIVIDUAL_FETCH_BATCH_SIZE]
-            batch_results = await asyncio.gather(*[fetch_one(x) for x in batch])
-            for key, config in batch_results:
-                if config is not None:
-                    configs[key] = config
-                    fetched_count += 1
-                else:
-                    failed_count += 1
-        return configs, failed_count, skipped_count
-
-    def _score_config_entries(
-        self,
-        scored: list[tuple[str, str, str | None, int]],
-        configs: dict[str, dict[str, Any]],
-        query_lower: str,
-        exact_match: bool,
-    ) -> list[dict[str, Any]]:
-        """Score each ``(entity_id, friendly_name, key, name_score)`` against its config.
-
-        Returns one raw match record per entry clearing its threshold. Each
-        per-type caller maps these records into its own result shape.
-        """
-        matches: list[dict[str, Any]] = []
-        for entity_id, friendly_name, key, name_score in scored:
-            config = configs.get(key, {}) if key else {}
-            config_match_score = (
-                self._search_in_dict(config, query_lower, exact_match) if config else 0
-            )
-            total_score, threshold, match_in_name = self._score_deep_match(
-                entity_id,
-                friendly_name,
-                name_score,
-                config_match_score,
-                query_lower,
-                exact_match,
-            )
-            if total_score >= threshold:
-                matches.append(
-                    {
-                        "entity_id": entity_id,
-                        "friendly_name": friendly_name,
-                        "key": key,
-                        "config": config,
-                        "score": total_score,
-                        "match_in_name": match_in_name,
-                        "match_in_config": config_match_score >= threshold,
-                    }
-                )
-        return matches
-
     async def _deep_search_automations(
         self,
         all_entities: list[dict[str, Any]],
@@ -523,253 +383,6 @@ class DeepSearchMixin(_SearchBase):
                 scored, configs, query_lower, exact_match
             )
         ]
-
-    async def _walk_scene_registry(
-        self, configs: dict[str, dict[str, Any]]
-    ) -> tuple[set[str], dict[str, str], bool]:
-        """Walk the entity registry once for scene metadata (Phase 2.5).
-
-        Returns ``(homeassistant_scene_uids, slug_to_storage_id, registry_failed)``
-        and mutates ``configs`` in place, aliasing each bulk-fetched config under
-        its entity-id slug. Two outputs:
-
-        1. ``homeassistant_scene_uids`` -- unique_ids backed by
-           ``platform == "homeassistant"`` (HA's storage collection).
-           Integration-managed scenes (Hue, IKEA, deCONZ, ...) are entity-only;
-           the per-id REST endpoint ``/config/scene/config/<id>`` can't fetch
-           them and treating their 404s as ``failed_count`` produces a
-           misleading ``partial: true`` flag (issue #1168 R3 blocker 2).
-        2. Slug-keyed aliases pointing at the bulk-fetched config. HA derives a
-           scene's entity_id from the ``name`` field via its own slugify
-           (collapsing runs of underscores, replacing all non-alnum with
-           underscores, etc.); approximating that with ``.replace()`` chains
-           produces near-misses.
-
-        Run unconditionally so the platform filter is available even when the
-        bulk fetch returned nothing (the common Hue-only case).
-        """
-        homeassistant_scene_uids: set[str] = set()
-        # Issue #1168 R7 blocker 17/21: registry-derived slug->storage map for
-        # the result-builder fallback, keeping the storage key correct for any
-        # scene the registry knows about regardless of bulk-fetch coverage.
-        slug_to_storage_id: dict[str, str] = {}
-        try:
-            reg_resp = await asyncio.wait_for(
-                self.client.send_websocket_message(
-                    {"type": "config/entity_registry/list"}
-                ),
-                timeout=BULK_WEBSOCKET_TIMEOUT,
-            )
-            if isinstance(reg_resp, dict) and reg_resp.get("success"):
-                for entry in reg_resp.get("result") or []:
-                    self._index_scene_registry_entry(
-                        entry, configs, homeassistant_scene_uids, slug_to_storage_id
-                    )
-        except Exception as e:
-            # Issue #1168 R5 blocker 11: promote DEBUG -> WARNING and signal the
-            # fallback so partial_reason can explain why the count looks
-            # elevated. A true registry outage previously looked identical to
-            # the steady-state happy path on stderr.
-            logger.warning(
-                "Scene entity-registry augmentation failed: %s; "
-                "integration-platform filter unavailable, attempting all scenes",
-                e,
-            )
-            return homeassistant_scene_uids, slug_to_storage_id, True
-        return homeassistant_scene_uids, slug_to_storage_id, False
-
-    @staticmethod
-    def _index_scene_registry_entry(
-        entry: dict[str, Any],
-        configs: dict[str, dict[str, Any]],
-        homeassistant_scene_uids: set[str],
-        slug_to_storage_id: dict[str, str],
-    ) -> None:
-        """Record one entity-registry scene entry into the registry-walk outputs."""
-        ent_id = entry.get("entity_id") or ""
-        uid = entry.get("unique_id")
-        if not ent_id.startswith("scene.") or not uid:
-            return
-        if entry.get("platform") == "homeassistant":
-            homeassistant_scene_uids.add(uid)
-        slug = ent_id.removeprefix("scene.")
-        if slug:
-            slug_to_storage_id[slug] = uid
-        if uid in configs and slug and slug != uid:
-            configs[slug] = configs[uid]
-
-    @staticmethod
-    def _select_scene_ids_to_fetch(
-        scored: list[tuple[str, str, str | None, int]],
-        configs: dict[str, dict[str, Any]],
-        homeassistant_scene_uids: set[str],
-    ) -> tuple[list[str], int]:
-        """Pick scene ids needing a per-id fetch, skipping integration-managed ones.
-
-        Issue #1168 R3 blocker 2: integration-managed scenes 404 on the per-id
-        REST endpoint by design, so surfacing those as fetch failures masks real
-        errors. They are counted separately (returned as ``integration_skipped``).
-        When the registry call failed (``homeassistant_scene_uids`` empty), fall
-        back to attempting all scenes -- false partials beat dropping legitimate
-        HA-managed scenes silently.
-
-        Returns ``(sids_to_fetch, integration_skipped_count)``.
-        """
-        if not homeassistant_scene_uids:
-            return [sid for _, _, sid, _ in scored if sid and sid not in configs], 0
-        sids: list[str] = []
-        integration_skipped = 0
-        for _, _, sid, _ in scored:
-            if not sid or sid in configs:
-                continue
-            if sid in homeassistant_scene_uids:
-                sids.append(sid)
-            else:
-                integration_skipped += 1
-        return sids, integration_skipped
-
-    @staticmethod
-    def _resolve_scene_storage_id(
-        scene_config: dict[str, Any],
-        scene_id: str | None,
-        slug_to_storage_id: dict[str, str],
-    ) -> str | None:
-        """Resolve a scene's storage key (the contract used by ha_config_*_scene).
-
-        Issue #1168 R6/R7 blockers 17/21: three-tier resolution:
-          1. ``scene_config["id"]`` -- present whenever the bulk fetch carried it.
-          2. ``slug_to_storage_id`` -- registry-derived; covers integration-
-             managed scenes and any scene whose bulk record omitted ``id``.
-          3. ``scene_id`` itself (the entity-id slug) -- final fallback when the
-             registry walk also failed; surfaced via ``logger.warning`` so the
-             silent-slug-mismatch path becomes observable.
-        """
-        config_id = scene_config.get("id") if isinstance(scene_config, dict) else None
-        if isinstance(config_id, str):
-            return config_id
-        if scene_id in slug_to_storage_id:
-            return slug_to_storage_id[scene_id]
-        logger.warning(
-            "ha_deep_search scene result fell back to entity-id slug for "
-            "scene_id=%r -- neither bulk config nor registry walk produced a "
-            "storage key. ``ha_config_get_scene`` will rely on its resolver "
-            "remap to land on the right scene.",
-            scene_id,
-        )
-        return scene_id
-
-    async def _deep_search_scenes(
-        self,
-        all_entities: list[dict[str, Any]],
-        query_lower: str,
-        exact_match: bool,
-    ) -> tuple[list[dict[str, Any]], int, int, int, bool]:
-        """Deep-search scenes: 3-tier strategy plus registry-walk augmentation.
-
-        Scenes have no listing primitive, so entities are enumerated from
-        get_states() and configs fetched per id. Returns the scene results plus
-        the four signals that drive the response ``partial`` flag:
-        ``(results, failed_count, skipped_count, integration_skipped, registry_failed)``.
-        """
-        scene_entities = [
-            e for e in all_entities if e.get("entity_id", "").startswith("scene.")
-        ]
-
-        # Phase 1: Score all scenes by name (instant)
-        scored: list[tuple[str, str, str | None, int]] = []
-        for entity in scene_entities:
-            entity_id = entity.get("entity_id", "")
-            friendly_name = entity.get("attributes", {}).get("friendly_name", entity_id)
-            scene_id = entity_id.replace("scene.", "")
-            name_score = self.fuzzy_searcher._calculate_entity_score(
-                entity_id, friendly_name, "scene", query_lower
-            )
-            scored.append((entity_id, friendly_name, scene_id, name_score))
-
-        # Phase 2: bulk fetch
-        configs = await self._bulk_fetch_configs(
-            "/config/scene/config",
-            ["config/scene/config/list", "scene/config/list"],
-            lambda item: (
-                item.get("id") or item.get("name", "").lower().replace(" ", "_")
-            ),
-            INDIVIDUAL_CONFIG_TIMEOUT,
-            "Scene",
-        )
-        bulk_fetched = configs is not None
-        if configs is None:
-            configs = {}
-
-        # Phase 2.5: registry walk (runs unconditionally, mutates ``configs``,
-        # and must precede Attempt C since the integration-skip filter depends
-        # on its homeassistant_scene_uids output).
-        (
-            homeassistant_scene_uids,
-            slug_to_storage_id,
-            registry_failed,
-        ) = await self._walk_scene_registry(configs)
-
-        failed_count = 0
-        skipped_count = 0
-        integration_skipped = 0
-
-        # Attempt C: parallel per-id fetch with a wall-clock budget so a few
-        # slow scenes don't tank the whole search.
-        if not bulk_fetched:
-            sids_to_fetch, integration_skipped = self._select_scene_ids_to_fetch(
-                scored, configs, homeassistant_scene_uids
-            )
-
-            async def _fetch_scene_config(
-                sid: str,
-            ) -> tuple[str, dict[str, Any] | None]:
-                try:
-                    config_resp = await asyncio.wait_for(
-                        self.client.get_scene_config(sid),
-                        timeout=INDIVIDUAL_CONFIG_TIMEOUT,
-                    )
-                    return (sid, config_resp.get("config", {}))
-                except Exception as e:
-                    logger.debug(f"Scene individual config fetch ({sid}) failed: {e}")
-                    return (sid, None)
-
-            (
-                fetched_configs,
-                failed_count,
-                skipped_count,
-            ) = await self._individual_fetch_budgeted(
-                sids_to_fetch,
-                _fetch_scene_config,
-                SCENE_CONFIG_TIME_BUDGET,
-                "Scene",
-                "scenes",
-            )
-            configs.update(fetched_configs)
-
-        # Phase 3: Score scenes, resolving each match's storage key
-        scene_results: list[dict[str, Any]] = []
-        for m in self._score_config_entries(scored, configs, query_lower, exact_match):
-            scene_config = m["config"]
-            scene_results.append(
-                {
-                    "entity_id": m["entity_id"],
-                    "scene_id": self._resolve_scene_storage_id(
-                        scene_config, m["key"], slug_to_storage_id
-                    ),
-                    "friendly_name": m["friendly_name"],
-                    "score": m["score"],
-                    "match_in_name": m["match_in_name"],
-                    "match_in_config": m["match_in_config"],
-                    "config": scene_config if scene_config else None,
-                }
-            )
-        return (
-            scene_results,
-            failed_count,
-            skipped_count,
-            integration_skipped,
-            registry_failed,
-        )
 
     async def _search_helper_type(
         self,
@@ -1018,47 +631,6 @@ class DeepSearchMixin(_SearchBase):
         self._apply_scene_partial_flag(response, scene_stats)
         return response
 
-    @staticmethod
-    def _apply_scene_partial_flag(
-        response: dict[str, Any], scene_stats: dict[str, Any]
-    ) -> None:
-        """Set ``partial``/``partial_reason`` from the scene Attempt-C signals.
-
-        Only set ``partial: True`` when something actually went wrong;
-        downstream consumers treat absence as success. Issue #1168 R3 blocker 2:
-        integration-managed scenes intentionally skip the per-id fetch and never
-        raise ``partial`` on their own (the count is informational).
-        """
-        failed = scene_stats["failed"]
-        skipped = scene_stats["skipped"]
-        if not (failed or skipped):
-            return
-        response["partial"] = True
-        reason_parts = [
-            f"Scene config fetch incomplete: {failed} failed, "
-            f"{skipped} skipped (time budget)."
-        ]
-        if scene_stats["integration_skipped"]:
-            reason_parts.append(
-                f" {scene_stats['integration_skipped']} integration-managed "
-                "scenes are scored by attribute only (no per-id fetch)."
-            )
-        if scene_stats["registry_failed"]:
-            # Issue #1168 R5 blocker 11: when the registry fetch errors, the
-            # integration-platform filter is unavailable and Attempt C falls
-            # back to attempting all scenes -- surface that so an elevated
-            # failed_count isn't mistaken for a real config outage.
-            reason_parts.append(
-                " Entity-registry fetch failed; integration-platform filter "
-                "unavailable, attempted all scenes (false-positive failures "
-                "expected for integration-managed scenes)."
-            )
-        reason_parts.append(
-            " Some scene matches may be missing config data; tune "
-            "HAMCP_SCENE_CONFIG_TIME_BUDGET to raise the budget."
-        )
-        response["partial_reason"] = "".join(reason_parts)
-
     async def _search_flow_helpers(
         self,
         query_lower: str,
@@ -1193,151 +765,3 @@ class DeepSearchMixin(_SearchBase):
         if include_config:
             result["config"] = options
         return result
-
-    def _score_deep_match(
-        self,
-        entity_id: str,
-        friendly_name: str,
-        fuzzy_name_score: int,
-        config_match_score: int,
-        query_lower: str,
-        exact_match: bool,
-    ) -> tuple[int, int, bool]:
-        """Compute total score, threshold, and match_in_name for a deep search result.
-
-        Returns (total_score, threshold, match_in_name).
-        """
-        if exact_match:
-            name_exact = (
-                100
-                if query_lower in entity_id.lower()
-                or query_lower in friendly_name.lower()
-                else 0
-            )
-            total_score = max(name_exact, config_match_score)
-            return total_score, 100, name_exact >= 100
-        else:
-            total_score = max(fuzzy_name_score, config_match_score)
-            threshold = self.settings.fuzzy_threshold
-            return total_score, threshold, fuzzy_name_score >= threshold
-
-    def _search_in_dict(
-        self,
-        data: dict[str, Any] | list[Any] | Any,
-        query: str,
-        exact_match: bool = False,
-    ) -> int:
-        """Search for query in nested dictionary/list structures.
-
-        When exact_match is True, uses substring matching (returns 100 if found, 0 if not).
-        When exact_match is False, collects all string leaves, tokenizes them into a
-        single BM25 document, and scores against the query tokens.  Falls back to
-        token-level SequenceMatcher if BM25 returns 0 (typo correction).
-        """
-        if exact_match:
-            return self._search_in_dict_exact(data, query)
-
-        # Fuzzy path: collect all string leaves, build a single tokenised document
-        leaves: list[str] = []
-        self._collect_string_leaves(data, leaves)
-        if not leaves:
-            return 0
-
-        query_tokens = tokenize(query)
-        if not query_tokens:
-            return 0
-
-        # Build a single flat token list from all leaves
-        doc_tokens: list[str] = []
-        for leaf in leaves:
-            doc_tokens.extend(tokenize(leaf))
-
-        if not doc_tokens:
-            return 0
-
-        # Use BM25 with a 1-document corpus (the config dict as a single doc)
-        scorer = BM25Scorer()
-        scorer.fit([doc_tokens])
-        raw = scorer.score(query_tokens, 0)
-
-        if raw > 0:
-            # Normalise against the theoretical max (sum of IDF per query
-            # token). With a 1-document corpus every token's IDF is identical
-            # (~0.288 with smoothing), so the ratio effectively measures how
-            # many query tokens the config contains. Cap at 100 for the edge
-            # case where high TF pushes raw above the sum-of-IDFs baseline.
-            max_possible = scorer.max_possible_score(query_tokens)
-            if max_possible > 0:
-                return min(100, round(raw / max_possible * 100))
-            logger.warning(
-                "BM25 scored > 0 but max_possible IDF is 0; "
-                "query_tokens=%s, doc_tokens_len=%d",
-                query_tokens,
-                len(doc_tokens),
-            )
-            return 100
-
-        # Tier-3 fallback: token-level SequenceMatcher for typos
-        logger.debug(
-            "BM25 returned 0 for query_tokens=%s; "
-            "falling back to SequenceMatcher typo scoring over %d unique tokens",
-            query_tokens,
-            len(set(doc_tokens)),
-        )
-        best = 0
-        for qt in query_tokens:
-            for dt in set(doc_tokens):
-                best = max(best, calculate_ratio(qt, dt))
-        return best if best >= 70 else 0
-
-    @staticmethod
-    def _collect_string_leaves(
-        data: dict[str, Any] | list[Any] | Any, out: list[str]
-    ) -> None:
-        """Recursively collect all string representations from nested data."""
-        if isinstance(data, dict):
-            for key, value in data.items():
-                out.append(str(key))
-                DeepSearchMixin._collect_string_leaves(value, out)
-        elif isinstance(data, list):
-            for item in data:
-                DeepSearchMixin._collect_string_leaves(item, out)
-        elif isinstance(data, str):
-            out.append(data)
-        elif data is not None:
-            out.append(str(data))
-
-    @classmethod
-    def _search_in_dict_exact(
-        cls,
-        data: dict[str, Any] | list[Any] | Any,
-        query: str,
-    ) -> int:
-        """Exact substring search in nested structures (returns 100 or 0)."""
-        if isinstance(data, dict):
-            return cls._exact_in_dict(data, query)
-        if isinstance(data, list):
-            return cls._exact_in_list(data, query)
-        if isinstance(data, str):
-            return 100 if query in data.lower() else 0
-        if data is not None:
-            return 100 if query in str(data).lower() else 0
-        return 0
-
-    @classmethod
-    def _exact_in_dict(cls, data: dict[str, Any], query: str) -> int:
-        """Exact-match scan over a dict's keys and recursively over its values."""
-        for key, value in data.items():
-            if query in str(key).lower():
-                return 100
-            if cls._search_in_dict_exact(value, query) >= 100:
-                return 100
-        return 0
-
-    @classmethod
-    def _exact_in_list(cls, data: list[Any], query: str) -> int:
-        """Exact-match scan recursively over a list's items."""
-        for item in data:
-            if cls._search_in_dict_exact(item, query) >= 100:
-                return 100
-        return 0
